perf: 统一缓存模型并优化仓库配置缓存
- 将 VirtualRepoCache 重构为通用的 RepositoryCache
- 新增 PackageCache 缓存包元数据，减少数据库查询
- 引入 CacheManager 统一管理四种缓存类型
- 缓存管理页面支持多缓存类型切换查看
- 普通仓库下载路径增加配置缓存，避免每次请求查库
- 仓库配置变更时自动失效相关缓存
- CacheService 默认内存上限从 10GB 降至 2GB 并可配置
- 新增 ParseSize 函数支持带单位的文件大小解析

feat: 仓库配置缓存
- ProxyRouter 和 RepoRouter 统一使用 RepositoryCache
- 包操作时自动失效 PackageCache

refactor: 缓存架构升级
- 定义 CacheProvider 统一接口
- 为 CacheService、RepositoryCache、PermissionCache、PackageCache 提供适配器
- 前端缓存管理页面支持按类型查看和统计

+++ internal/model/package.go
@@ -71,7 +71,7 @@ type PackageVersion struct {
 	PublishedAt    time.Time           `gorm:"autoCreateTime" json:"published_at"`
 	PublishedBy    uint                `json:"published_by"`
 	Metadata       string              `gorm:"type:text" json:"metadata,omitempty"`
-	DownloadCount  int                 `gorm:"default:0" json:"download_count"`
+	DownloadCount  int64               `gorm:"default:0" json:"download_count"`
 	SizeBytes      int64               `gorm:"default:0" json:"size_bytes"`
 	ChecksumMD5    string              `gorm:"size:32" json:"checksum_md5,omitempty"`
 	ChecksumSHA256 string              `gorm:"size:64" json:"checksum_sha256,omitempty"`
@@ -92,7 +92,7 @@ type PackageFile struct {
 	SizeBytes      int64           `gorm:"default:0" json:"size_bytes"`
 	ChecksumSHA256 string          `gorm:"size:64" json:"checksum_sha256,omitempty"`
 	ChecksumMD5    string          `gorm:"size:32" json:"checksum_md5,omitempty"`
-	DownloadCount  int             `gorm:"default:0" json:"download_count"`
+	DownloadCount  int64           `gorm:"default:0" json:"download_count"`
 	Version        PackageVersion  `gorm:"foreignKey:VersionID" json:"-"`
 }
